tests/wpt/web-platform-tests/notifications/event-onerror-default-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Notifications Test: notification - onerror (basic)</title>
<link rel="author" title="Apple Inc." href="http://www.apple.com/">
<link rel="help" title="5 API" href="http://www.w3.org/TR/notifications/">
<meta name="flags" content="">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <div id="log"></div>
  <script>
    if (Notification.permission != "default") {
      document.getElementById("log").innerHTML = "Please clear your notification settings for web-platform.test and retry.";
    } else {
      t = async_test("the notification.onerror event can be invoked", {timeout: 50000});
    var notification = new Notification("New Email Received", {
      body: "Room 101"
    });
    notification.onerror = t.step_func(function(e) {
      assert_equals(Object.prototype.toString.call(e), "[object Event]", "the type of event");
      assert_equals(e.type, "error", "The event is the expected type.")
      t.done();
    });
    }
  </script>
</body>
